requirements.txt
Django
djangorestframework
google-generativeai
python-decouple
streamlit
